requirements.txt
# Automate web browser interaction from Python
selenium==3.141.0

# Allows you to send HTTP/1.1 requests
requests==2.28.1

# Extensions to the standard Python datetime module
python-dateutil==2.8.2
